Fix bug in skal-master: process all queued messages
This is a subtle bug: when rapidly sending messages to skal-master from
within the process, it is possible to queue 2 or more messages before
skal-master can process them. In such a case, the pipe we use to wake
skal-master up can accumulate more than one character, and we then need
to process all of them separately (there will be one character per
message to process).

Symptom was a random (but quite rare) deadlock when running skal-post
without expecting a response. This symptom does not occur anymore after
the fix.