skal-net: Explicitely handle EPIPE when sending
This is to inform the upper layer that the peer has closed the
connection while we were trying to send data to it. If we try to send a
lot of data, there is a significant chance that the send will fail with
that error number if the peer closes its socket.